códigos/Autoencoder/requirements.txt
torch==2.7.0
torchvision
numpy
matplotlib
Pillow
opencv-python
